GPDMA/SPI based updates of NeoPixels now working
This commit is a checkpoint where the code is now able to use the
GPDMA/SPI peripherals on the LPC17xx to automatically send out the
required single wire waveform for the WS2812 based NeoPixels. The CPU
is only used to convert the desired LED RGB values to a bitstream that
can be sent out over the SPI MOSI pin to generate the expected WS2812
waveform. This saves CPU cycles to be used for running animations.

The output from this test is a static pattern to just make sure that
the correct LEDs get set to the correct colour. Animations will be
added in future commits.