Add back buffer support
Previously the main thread would wait for the current front buffer to
be sent and then it would fill that front buffer in while the other was
being sent. This caused the CPU on average to wait half of a frame
period each time it called NeoPixel::set().

This new version has the main thread just filling in a separate
back buffer. That back buffer will then be copied into the next front
buffer to become available. This means the main thread only needs to
wait if the previous backbuffer update hasn't yet been copied into the
front buffer(s). The copy of the back buffer to the front buffer is
started in the interrupt handler that is triggered when a front buffer
has just finished being sent to the NeoPixel strip. The back buffer to
front buffer copy also takes advantage of the GPDMA peripheral on the
LPC17xx. This means that the CPU isn't utilized for this copy
operation and is free to run animation code and the conversion of RGB
pixel data to a format that can be sent over SPI.

This commit also includes these updates:
* The NeoPixel class now tracks the number of times the front buffer
  has been sent to the NeoPixels and the number of times that
  NeoPixel::set() has been called by the application code.
* main() uses these counters to display the maximum possible frame rate
  (front buffer flips) and the actual frame rate (::set calls).
* The makefile was updated since main.cpp no longer requires float
  support from printf().
* I added a DmaMemCopy() function to the GPDMA module. This function
  will reserve the lowest priority GPDMA channel for the memory copy
  and use it to copy the memory contents asynchronously. It will let
  the application know when the copy completes via a caller supplied
  callback. The current implementation doesn't queue up async memory
  copies and it will instead it will fall-back to using the CPU for the
  copy if there is already an async copy in progress.
* I fixed the DMA_IRQHandler() in GPDMA.c to not exit its loop until
  all of the currently signalled GPDMA interrupts have been handled by
  the handlers. Before it would stop even if a handler only handled a
  subset of the currently active interrupts.
* The dmaHeap*Alloc() routines now keeps the pointers it returns 8-byte
  aligned like the main newlib heap.