Make rotary encoder code more robust
This update was inspired by the following code from the Adafruit team:
https://github.com/adafruit/Adafruit-Trinket-USB/blob/master/TrinketHidCombo/examples/TrinketVolumeKnob/TrinketVolumeKnob.ino

It tracks the history of encoder state transitions that are seen
between detent states. Once the encoder gets back to the detent state,
the history can be queried to determine if it was more likely to have
been rotated in a clockwise or counter clockwise direction.