Rewrite of rotary encoder filter
I rewrote the rotary encoder code so that it first filters each encoder
signal wire by only looking for pairs of clean HIGH (switch open)
pulses and it then infers the length of the noisier LOW (switch closed)
pulses. It returns the inferred LOW pulse and detected HIGH pulse back
to the caller in a queue of length 2 containing the corresponding 2
timestamped pulse descriptions.

The higher level code walks the queues returned for each of its signal
wires (A & B) to determine the direction of rotation. This works much
better than the previous version of the code since the code just infers
the LOW pulse (with all of its switch bounce) instead of trying to
detect it directly.