Create new LED holder PCB
The new PCB is rectangular and will extend horizontally from the end of
the branch wires. The LED will be bent at 90 degrees after being
soldered into this board so that it lays horizontally as well. I will
finalize the design once I have the flat cable in hand (currently on
order).